requirements.txt
streamlit
pdfplumber
sentence-transformers
faiss-cpu
transformers
requests
beautifulsoup4
openai
scikit-learn
google-generativeai
